Scrap AdvancedSettings, return to one page
