+++ public/sw.js
@@ -1,2 +1,2 @@
-if(!self.define){let e,s={};const a=(a,n)=>(a=new URL(a+".js",n).href,s[a]||new Promise((s=>{if("document"in self){const e=document.createElement("script");e.src=a,e.onload=s,document.head.appendChild(e)}else e=a,importScripts(a),s()})).then((()=>{let e=s[a];if(!e)throw new Error(`Module ${a} didn’t register its module`);return e})));self.define=(n,c)=>{const i=e||("document"in self?document.currentScript.src:"")||location.href;if(s[i])return;let r={};const t=e=>a(e,i),d={module:{uri:i},exports:r,require:t};s[i]=Promise.all(n.map((e=>d[e]||t(e)))).then((e=>(c(...e),r)))}}define(["./workbox-6316bd60"],(function(e){"use strict";importScripts("fallback-pfuVIl3SW_b_7o4TUTuW8.js"),self.addEventListener("message",(e=>{e.data&&"SKIP_WAITING"===e.data.type&&self.skipWaiting()})),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/static/chunks/592-b4d1ffd2bdb57917.js",revision:"b4d1ffd2bdb57917"},{url:"/_next/static/chunks/592-b4d1ffd2bdb57917.js.map",revision:"f11ff9806beb15873f932cc70dc25627"},{url:"/_next/static/chunks/721-6c370ff3d7a9bc9b.js",revision:"6c370ff3d7a9bc9b"},{url:"/_next/static/chunks/721-6c370ff3d7a9bc9b.js.map",revision:"ae324f097ca648a4380cf6e12961088e"},{url:"/_next/static/chunks/788-232aa896862420bc.js",revision:"232aa896862420bc"},{url:"/_next/static/chunks/788-232aa896862420bc.js.map",revision:"87c063f68c5e98657b81769ce0d36b2e"},{url:"/_next/static/chunks/framework-da8ad0d3df154e7f.js",revision:"da8ad0d3df154e7f"},{url:"/_next/static/chunks/framework-da8ad0d3df154e7f.js.map",revision:"e43122222b79391baf565e0b2cea680d"},{url:"/_next/static/chunks/main-dda75576ff8dcbb9.js",revision:"dda75576ff8dcbb9"},{url:"/_next/static/chunks/main-dda75576ff8dcbb9.js.map",revision:"8ea4d1b06bf6684eb000fbfab9a361f1"},{url:"/_next/static/chunks/pages/404-eaa4bffbb7035117.js",revision:"eaa4bffbb7035117"},{url:"/_next/static/chunks/pages/404-eaa4bffbb7035117.js.map",revision:"f37058f112458a9e5119d78fe3530166"},{url:"/_next/static/chunks/pages/500-c4cbe2f5aecfffec.js",revision:"c4cbe2f5aecfffec"},{url:"/_next/static/chunks/pages/500-c4cbe2f5aecfffec.js.map",revision:"c99339bcaebcb0b308cb94a5f021a5a0"},{url:"/_next/static/chunks/pages/_app-ba7a1f07b5f01357.js",revision:"ba7a1f07b5f01357"},{url:"/_next/static/chunks/pages/_app-ba7a1f07b5f01357.js.map",revision:"5b1468275dec56b2978b54ecc07c4475"},{url:"/_next/static/chunks/pages/_error-e2f2bfcd5131d4b7.js",revision:"e2f2bfcd5131d4b7"},{url:"/_next/static/chunks/pages/_error-e2f2bfcd5131d4b7.js.map",revision:"304eea13ca7ffb487de091a049524fe8"},{url:"/_next/static/chunks/pages/_offline-9581b9c9c412baa8.js",revision:"9581b9c9c412baa8"},{url:"/_next/static/chunks/pages/_offline-9581b9c9c412baa8.js.map",revision:"e85d74474a2c63d9555bf5a50252928e"},{url:"/_next/static/chunks/pages/about/%5Bslug%5D-c3fffc358bfd4d3f.js",revision:"c3fffc358bfd4d3f"},{url:"/_next/static/chunks/pages/about/%5Bslug%5D-c3fffc358bfd4d3f.js.map",revision:"e7d6eef19d11fdb645feec3c1581c0e1"},{url:"/_next/static/chunks/pages/browse/%5Bslug%5D-09fbf370a0741f41.js",revision:"09fbf370a0741f41"},{url:"/_next/static/chunks/pages/browse/%5Bslug%5D-09fbf370a0741f41.js.map",revision:"156d0b6944eb27f1fdd13b69339182ed"},{url:"/_next/static/chunks/pages/commit-837588f7d4edcf5b.js",revision:"837588f7d4edcf5b"},{url:"/_next/static/chunks/pages/commit-837588f7d4edcf5b.js.map",revision:"5de01ad021835b7d3f17b1058fad3dcc"},{url:"/_next/static/chunks/pages/docs/%5Bslug%5D-83327271cd2eca56.js",revision:"83327271cd2eca56"},{url:"/_next/static/chunks/pages/docs/%5Bslug%5D-83327271cd2eca56.js.map",revision:"95ab1323c1353d7eca6e60b1bee8c623"},{url:"/_next/static/chunks/pages/index-d92bd74542f86a9c.js",revision:"d92bd74542f86a9c"},{url:"/_next/static/chunks/pages/index-d92bd74542f86a9c.js.map",revision:"72bccfaf6965e3e3406d5ac5a0857c6d"},{url:"/_next/static/chunks/pages/settings-315ad6b8563b83b6.js",revision:"315ad6b8563b83b6"},{url:"/_next/static/chunks/pages/settings-315ad6b8563b83b6.js.map",revision:"dd474f1a3937d3e2c3897129fcb5fae8"},{url:"/_next/static/chunks/polyfills-5cd94c89d3acac5f.js",revision:"99442aec5788bccac9b2f0ead2afdd6b"},{url:"/_next/static/chunks/webpack-70094112286a753f.js",revision:"70094112286a753f"},{url:"/_next/static/chunks/webpack-70094112286a753f.js.map",revision:"18c31cc0ab505ab4428f85db0cf9b6a8"},{url:"/_next/static/css/57ef369b041b297e.css",revision:"57ef369b041b297e"},{url:"/_next/static/css/57ef369b041b297e.css.map",revision:"7ab35e46206052e77e74dd54901610ee"},{url:"/_next/static/media/atkinson-hyperlegible-all-400-normal.a9b243f0.woff",revision:"a9b243f0"},{url:"/_next/static/media/atkinson-hyperlegible-latin-400-normal.a1159474.woff2",revision:"a1159474"},{url:"/_next/static/media/atkinson-hyperlegible-latin-ext-400-normal.98b7be61.woff2",revision:"98b7be61"},{url:"/_next/static/media/public-sans-all-400-normal.ae05de44.woff",revision:"ae05de44"},{url:"/_next/static/media/public-sans-all-600-normal.07428ec5.woff",revision:"07428ec5"},{url:"/_next/static/media/public-sans-latin-400-normal.66c4eba4.woff2",revision:"66c4eba4"},{url:"/_next/static/media/public-sans-latin-600-normal.7e06741e.woff2",revision:"7e06741e"},{url:"/_next/static/media/public-sans-latin-ext-400-normal.fd422b1b.woff2",revision:"fd422b1b"},{url:"/_next/static/media/public-sans-latin-ext-600-normal.47b27dc8.woff2",revision:"47b27dc8"},{url:"/_next/static/pfuVIl3SW_b_7o4TUTuW8/_buildManifest.js",revision:"8293e67bd0f79b10a77c7c65014e3550"},{url:"/_next/static/pfuVIl3SW_b_7o4TUTuW8/_middlewareManifest.js",revision:"fb2823d66b3e778e04a3f681d0d2fb19"},{url:"/_next/static/pfuVIl3SW_b_7o4TUTuW8/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_offline",revision:"pfuVIl3SW_b_7o4TUTuW8"},{url:"/brand-assets/apple-icon-180.png",revision:"8780940006caca5b1e373f06e82fd1af"},{url:"/brand-assets/apple-splash-1125-2436.jpg",revision:"86fbd1cb121ceaf3e08bdcebba56c7be"},{url:"/brand-assets/apple-splash-1136-640.jpg",revision:"ede7a029c432b282e2cc14fe00239672"},{url:"/brand-assets/apple-splash-1170-2532.jpg",revision:"91d14a1e49019231cc28c998a7f698fd"},{url:"/brand-assets/apple-splash-1242-2208.jpg",revision:"31eeaa7d5bf26a403fb3a16774d6483e"},{url:"/brand-assets/apple-splash-1242-2688.jpg",revision:"7b495b042942bda0aa365bb7dbed8b3b"},{url:"/brand-assets/apple-splash-1284-2778.jpg",revision:"f87029e087d3020ffd0f35447c0fdf14"},{url:"/brand-assets/apple-splash-1334-750.jpg",revision:"165246c883305f86c45fd110818473bb"},{url:"/brand-assets/apple-splash-1536-2048.jpg",revision:"8d3dadbd41be1abac9bcbea2371dfd82"},{url:"/brand-assets/apple-splash-1620-2160.jpg",revision:"4ce8e458e1aa5fac5fb222fa7cb8cea9"},{url:"/brand-assets/apple-splash-1668-2224.jpg",revision:"5878b1a40b476a5364b562956d82f2c1"},{url:"/brand-assets/apple-splash-1668-2388.jpg",revision:"dfb78f53e7f2f292ec6f21639309663b"},{url:"/brand-assets/apple-splash-1792-828.jpg",revision:"c227af436352a6caaf93acb7f72c5c33"},{url:"/brand-assets/apple-splash-2048-1536.jpg",revision:"c4774c952cfb7d92873c477d3476de1a"},{url:"/brand-assets/apple-splash-2048-2732.jpg",revision:"8c8164473e2c039346bab0f04eb911e3"},{url:"/brand-assets/apple-splash-2160-1620.jpg",revision:"c36f5e613ea4fac80a8899354c85902e"},{url:"/brand-assets/apple-splash-2208-1242.jpg",revision:"275493f4cbab6255101666c01ee72b8c"},{url:"/brand-assets/apple-splash-2224-1668.jpg",revision:"1300f65f70553e9392a80973ed272474"},{url:"/brand-assets/apple-splash-2388-1668.jpg",revision:"f2090ccbd3b7562ac39439f723c7136f"},{url:"/brand-assets/apple-splash-2436-1125.jpg",revision:"d2f5d6ec44eb0e0c8c4e0cedd690bd05"},{url:"/brand-assets/apple-splash-2532-1170.jpg",revision:"efac034a67fbc6cc14f8f35a12a3456a"},{url:"/brand-assets/apple-splash-2688-1242.jpg",revision:"7c90bdff925d7fcf9ab580a17d295d6a"},{url:"/brand-assets/apple-splash-2732-2048.jpg",revision:"516acb6571386fdf76de28686f937e87"},{url:"/brand-assets/apple-splash-2778-1284.jpg",revision:"efe493e5a9fc277a380c209b9e97d1c9"},{url:"/brand-assets/apple-splash-640-1136.jpg",revision:"069cd3496edcac61154a939ae02bb835"},{url:"/brand-assets/apple-splash-750-1334.jpg",revision:"e106f8b20677b8d7df0f0e2dc61824c3"},{url:"/brand-assets/apple-splash-828-1792.jpg",revision:"296bdce7a826fa315ad7e3cb115ee0c8"},{url:"/brand-assets/favicon-196.png",revision:"1405dff5bbadc38724b93ad4c0a8f23f"},{url:"/brand-assets/manifest-icon-192.png",revision:"e4312fb3f6e55c8b7e1070bdc2522f52"},{url:"/brand-assets/manifest-icon-512.png",revision:"40a35280399add43df9c56753b332bbe"},{url:"/manifest.json",revision:"53149ae646cff0058e467800c1a5be88"},{url:"/markdown/about/privacy.mdx",revision:"e8add1238d2aa01184ad7dfdb52cef08"},{url:"/markdown/about/terms.mdx",revision:"7df0a04cf2ef2a9b794bac9a3336d095"},{url:"/markdown/browse/demo.mdx",revision:"11f850a9fc83b08a86c6a93cc96dc934"},{url:"/markdown/docs/accessibility.mdx",revision:"724a2f0290f63502e0ceaf655780c6cf"},{url:"/markdown/docs/errors.mdx",revision:"f3dfb09ebb36f71a4b7c9844456d0bbe"},{url:"/markdown/docs/getting-started.mdx",revision:"a509320283f7d88387afc658990f81b2"},{url:"/markdown/docs/introduction.mdx",revision:"59f32c7b17533d9eaa4473fa48a5cac5"},{url:"/markdown/docs/keyboard-shortcuts.mdx",revision:"a7a271c9daf699e16be40c7812fae65f"},{url:"/markdown/docs/settings.mdx",revision:"23ca6ddf6cf2dc7bec50df099f754f4d"},{url:"/markdown/docs/sharing.mdx",revision:"17fbe212a9930751ea3671a20e12522a"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:a,state:n})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s},{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;const s=e.pathname;return!s.startsWith("/api/auth/")&&!!s.startsWith("/api/")}),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")}),new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute((({url:e})=>!(self.origin===e.origin)),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET")}));
+if(!self.define){let e,s={};const a=(a,n)=>(a=new URL(a+".js",n).href,s[a]||new Promise((s=>{if("document"in self){const e=document.createElement("script");e.src=a,e.onload=s,document.head.appendChild(e)}else e=a,importScripts(a),s()})).then((()=>{let e=s[a];if(!e)throw new Error(`Module ${a} didn’t register its module`);return e})));self.define=(n,c)=>{const i=e||("document"in self?document.currentScript.src:"")||location.href;if(s[i])return;let r={};const t=e=>a(e,i),d={module:{uri:i},exports:r,require:t};s[i]=Promise.all(n.map((e=>d[e]||t(e)))).then((e=>(c(...e),r)))}}define(["./workbox-6316bd60"],(function(e){"use strict";importScripts("fallback-zEJTFaScQk_0o-0JOC8ig.js"),self.addEventListener("message",(e=>{e.data&&"SKIP_WAITING"===e.data.type&&self.skipWaiting()})),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/static/chunks/592-b4d1ffd2bdb57917.js",revision:"b4d1ffd2bdb57917"},{url:"/_next/static/chunks/592-b4d1ffd2bdb57917.js.map",revision:"f11ff9806beb15873f932cc70dc25627"},{url:"/_next/static/chunks/721-b9b5cff67e3de81b.js",revision:"b9b5cff67e3de81b"},{url:"/_next/static/chunks/721-b9b5cff67e3de81b.js.map",revision:"b010a7f5fb48efb1b68081994bff82c7"},{url:"/_next/static/chunks/788-5b316538cfd64310.js",revision:"5b316538cfd64310"},{url:"/_next/static/chunks/788-5b316538cfd64310.js.map",revision:"4c830f5936808c5089ebc5e9edbd9f8b"},{url:"/_next/static/chunks/framework-da8ad0d3df154e7f.js",revision:"da8ad0d3df154e7f"},{url:"/_next/static/chunks/framework-da8ad0d3df154e7f.js.map",revision:"e43122222b79391baf565e0b2cea680d"},{url:"/_next/static/chunks/main-dda75576ff8dcbb9.js",revision:"dda75576ff8dcbb9"},{url:"/_next/static/chunks/main-dda75576ff8dcbb9.js.map",revision:"8ea4d1b06bf6684eb000fbfab9a361f1"},{url:"/_next/static/chunks/pages/404-eaa4bffbb7035117.js",revision:"eaa4bffbb7035117"},{url:"/_next/static/chunks/pages/404-eaa4bffbb7035117.js.map",revision:"f37058f112458a9e5119d78fe3530166"},{url:"/_next/static/chunks/pages/500-c4cbe2f5aecfffec.js",revision:"c4cbe2f5aecfffec"},{url:"/_next/static/chunks/pages/500-c4cbe2f5aecfffec.js.map",revision:"c99339bcaebcb0b308cb94a5f021a5a0"},{url:"/_next/static/chunks/pages/_app-ba7a1f07b5f01357.js",revision:"ba7a1f07b5f01357"},{url:"/_next/static/chunks/pages/_app-ba7a1f07b5f01357.js.map",revision:"5b1468275dec56b2978b54ecc07c4475"},{url:"/_next/static/chunks/pages/_error-e2f2bfcd5131d4b7.js",revision:"e2f2bfcd5131d4b7"},{url:"/_next/static/chunks/pages/_error-e2f2bfcd5131d4b7.js.map",revision:"304eea13ca7ffb487de091a049524fe8"},{url:"/_next/static/chunks/pages/_offline-9581b9c9c412baa8.js",revision:"9581b9c9c412baa8"},{url:"/_next/static/chunks/pages/_offline-9581b9c9c412baa8.js.map",revision:"e85d74474a2c63d9555bf5a50252928e"},{url:"/_next/static/chunks/pages/about/%5Bslug%5D-c3fffc358bfd4d3f.js",revision:"c3fffc358bfd4d3f"},{url:"/_next/static/chunks/pages/about/%5Bslug%5D-c3fffc358bfd4d3f.js.map",revision:"e7d6eef19d11fdb645feec3c1581c0e1"},{url:"/_next/static/chunks/pages/browse/%5Bslug%5D-09fbf370a0741f41.js",revision:"09fbf370a0741f41"},{url:"/_next/static/chunks/pages/browse/%5Bslug%5D-09fbf370a0741f41.js.map",revision:"156d0b6944eb27f1fdd13b69339182ed"},{url:"/_next/static/chunks/pages/commit-837588f7d4edcf5b.js",revision:"837588f7d4edcf5b"},{url:"/_next/static/chunks/pages/commit-837588f7d4edcf5b.js.map",revision:"5de01ad021835b7d3f17b1058fad3dcc"},{url:"/_next/static/chunks/pages/docs/%5Bslug%5D-83327271cd2eca56.js",revision:"83327271cd2eca56"},{url:"/_next/static/chunks/pages/docs/%5Bslug%5D-83327271cd2eca56.js.map",revision:"95ab1323c1353d7eca6e60b1bee8c623"},{url:"/_next/static/chunks/pages/index-d92bd74542f86a9c.js",revision:"d92bd74542f86a9c"},{url:"/_next/static/chunks/pages/index-d92bd74542f86a9c.js.map",revision:"72bccfaf6965e3e3406d5ac5a0857c6d"},{url:"/_next/static/chunks/pages/settings-cc8a9c6a72784dfd.js",revision:"cc8a9c6a72784dfd"},{url:"/_next/static/chunks/pages/settings-cc8a9c6a72784dfd.js.map",revision:"6fbf1610530f47ed60db64c7debf7899"},{url:"/_next/static/chunks/polyfills-5cd94c89d3acac5f.js",revision:"99442aec5788bccac9b2f0ead2afdd6b"},{url:"/_next/static/chunks/webpack-70094112286a753f.js",revision:"70094112286a753f"},{url:"/_next/static/chunks/webpack-70094112286a753f.js.map",revision:"18c31cc0ab505ab4428f85db0cf9b6a8"},{url:"/_next/static/css/57ef369b041b297e.css",revision:"57ef369b041b297e"},{url:"/_next/static/css/57ef369b041b297e.css.map",revision:"7ab35e46206052e77e74dd54901610ee"},{url:"/_next/static/media/atkinson-hyperlegible-all-400-normal.a9b243f0.woff",revision:"a9b243f0"},{url:"/_next/static/media/atkinson-hyperlegible-latin-400-normal.a1159474.woff2",revision:"a1159474"},{url:"/_next/static/media/atkinson-hyperlegible-latin-ext-400-normal.98b7be61.woff2",revision:"98b7be61"},{url:"/_next/static/media/public-sans-all-400-normal.ae05de44.woff",revision:"ae05de44"},{url:"/_next/static/media/public-sans-all-600-normal.07428ec5.woff",revision:"07428ec5"},{url:"/_next/static/media/public-sans-latin-400-normal.66c4eba4.woff2",revision:"66c4eba4"},{url:"/_next/static/media/public-sans-latin-600-normal.7e06741e.woff2",revision:"7e06741e"},{url:"/_next/static/media/public-sans-latin-ext-400-normal.fd422b1b.woff2",revision:"fd422b1b"},{url:"/_next/static/media/public-sans-latin-ext-600-normal.47b27dc8.woff2",revision:"47b27dc8"},{url:"/_next/static/zEJTFaScQk_0o-0JOC8ig/_buildManifest.js",revision:"bf55c5c709120394a84bc4937674d188"},{url:"/_next/static/zEJTFaScQk_0o-0JOC8ig/_middlewareManifest.js",revision:"fb2823d66b3e778e04a3f681d0d2fb19"},{url:"/_next/static/zEJTFaScQk_0o-0JOC8ig/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_offline",revision:"zEJTFaScQk_0o-0JOC8ig"},{url:"/brand-assets/apple-icon-180.png",revision:"8780940006caca5b1e373f06e82fd1af"},{url:"/brand-assets/apple-splash-1125-2436.jpg",revision:"86fbd1cb121ceaf3e08bdcebba56c7be"},{url:"/brand-assets/apple-splash-1136-640.jpg",revision:"ede7a029c432b282e2cc14fe00239672"},{url:"/brand-assets/apple-splash-1170-2532.jpg",revision:"91d14a1e49019231cc28c998a7f698fd"},{url:"/brand-assets/apple-splash-1242-2208.jpg",revision:"31eeaa7d5bf26a403fb3a16774d6483e"},{url:"/brand-assets/apple-splash-1242-2688.jpg",revision:"7b495b042942bda0aa365bb7dbed8b3b"},{url:"/brand-assets/apple-splash-1284-2778.jpg",revision:"f87029e087d3020ffd0f35447c0fdf14"},{url:"/brand-assets/apple-splash-1334-750.jpg",revision:"165246c883305f86c45fd110818473bb"},{url:"/brand-assets/apple-splash-1536-2048.jpg",revision:"8d3dadbd41be1abac9bcbea2371dfd82"},{url:"/brand-assets/apple-splash-1620-2160.jpg",revision:"4ce8e458e1aa5fac5fb222fa7cb8cea9"},{url:"/brand-assets/apple-splash-1668-2224.jpg",revision:"5878b1a40b476a5364b562956d82f2c1"},{url:"/brand-assets/apple-splash-1668-2388.jpg",revision:"dfb78f53e7f2f292ec6f21639309663b"},{url:"/brand-assets/apple-splash-1792-828.jpg",revision:"c227af436352a6caaf93acb7f72c5c33"},{url:"/brand-assets/apple-splash-2048-1536.jpg",revision:"c4774c952cfb7d92873c477d3476de1a"},{url:"/brand-assets/apple-splash-2048-2732.jpg",revision:"8c8164473e2c039346bab0f04eb911e3"},{url:"/brand-assets/apple-splash-2160-1620.jpg",revision:"c36f5e613ea4fac80a8899354c85902e"},{url:"/brand-assets/apple-splash-2208-1242.jpg",revision:"275493f4cbab6255101666c01ee72b8c"},{url:"/brand-assets/apple-splash-2224-1668.jpg",revision:"1300f65f70553e9392a80973ed272474"},{url:"/brand-assets/apple-splash-2388-1668.jpg",revision:"f2090ccbd3b7562ac39439f723c7136f"},{url:"/brand-assets/apple-splash-2436-1125.jpg",revision:"d2f5d6ec44eb0e0c8c4e0cedd690bd05"},{url:"/brand-assets/apple-splash-2532-1170.jpg",revision:"efac034a67fbc6cc14f8f35a12a3456a"},{url:"/brand-assets/apple-splash-2688-1242.jpg",revision:"7c90bdff925d7fcf9ab580a17d295d6a"},{url:"/brand-assets/apple-splash-2732-2048.jpg",revision:"516acb6571386fdf76de28686f937e87"},{url:"/brand-assets/apple-splash-2778-1284.jpg",revision:"efe493e5a9fc277a380c209b9e97d1c9"},{url:"/brand-assets/apple-splash-640-1136.jpg",revision:"069cd3496edcac61154a939ae02bb835"},{url:"/brand-assets/apple-splash-750-1334.jpg",revision:"e106f8b20677b8d7df0f0e2dc61824c3"},{url:"/brand-assets/apple-splash-828-1792.jpg",revision:"296bdce7a826fa315ad7e3cb115ee0c8"},{url:"/brand-assets/favicon-196.png",revision:"1405dff5bbadc38724b93ad4c0a8f23f"},{url:"/brand-assets/manifest-icon-192.png",revision:"e4312fb3f6e55c8b7e1070bdc2522f52"},{url:"/brand-assets/manifest-icon-512.png",revision:"40a35280399add43df9c56753b332bbe"},{url:"/manifest.json",revision:"53149ae646cff0058e467800c1a5be88"},{url:"/markdown/about/privacy.mdx",revision:"e8add1238d2aa01184ad7dfdb52cef08"},{url:"/markdown/about/terms.mdx",revision:"7df0a04cf2ef2a9b794bac9a3336d095"},{url:"/markdown/browse/demo.mdx",revision:"11f850a9fc83b08a86c6a93cc96dc934"},{url:"/markdown/docs/accessibility.mdx",revision:"724a2f0290f63502e0ceaf655780c6cf"},{url:"/markdown/docs/errors.mdx",revision:"f3dfb09ebb36f71a4b7c9844456d0bbe"},{url:"/markdown/docs/getting-started.mdx",revision:"a509320283f7d88387afc658990f81b2"},{url:"/markdown/docs/introduction.mdx",revision:"59f32c7b17533d9eaa4473fa48a5cac5"},{url:"/markdown/docs/keyboard-shortcuts.mdx",revision:"a7a271c9daf699e16be40c7812fae65f"},{url:"/markdown/docs/settings.mdx",revision:"23ca6ddf6cf2dc7bec50df099f754f4d"},{url:"/markdown/docs/sharing.mdx",revision:"17fbe212a9930751ea3671a20e12522a"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:a,state:n})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s},{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;const s=e.pathname;return!s.startsWith("/api/auth/")&&!!s.startsWith("/api/")}),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")}),new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute((({url:e})=>!(self.origin===e.origin)),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET")}));
 //# sourceMappingURL=sw.js.map
